Refactor Left Navigation Menu into It's Own Component
